Updated validation function to display messages or throw assertion error. Added some comments about what the code does.

+++ Utility/brush_init.m
@@ -48,7 +48,7 @@ function [model_input] = brush_init(numBrushes, isRolling, fs_sim, fs_save, t_in
         % data_range_2 = 0.1 / 3; % 10m/s; 36 km/h
         % data_range_3 = 0.2 / 3; % 20 m/s; 72 km/h
         smoothStep = smootherstep(edge0, edge1, t_save) .* (1 - smootherstep(edge2, edge3, t_save));
-        model_input.omega(:, 1) = data_range * smoothStep(:);
+        model_input.omega(:, 1) = single(data_range * smoothStep(:));
         % model_input.omega(:, 2) =  model_input.omega(:, 1) * 5;
         % model_input.omega(:, 3) =  model_input.omega(:, 2) * 2; linspace(0, 1, length(model_input.omega)).' .* 
         
@@ -58,8 +58,8 @@ function [model_input] = brush_init(numBrushes, isRolling, fs_sim, fs_save, t_in
         % Extend ramp downwards
         ramp = [ramp(1:end-1), fliplr(ramp)];
 
-        model_input.SR = ramp(:);
-        model_input.v0 =  model_input.omega * model_input.re ./ (model_input.SR + 1);
+        model_input.SR = single(ramp(:));
+        model_input.v0 =  single(model_input.omega * model_input.re ./ (model_input.SR + 1));
     else
         edge0 = 0.5;
         edge1 = 10.5;%25.5;
